+++ src/interfaces/user.ts
@@ -1,4 +1,5 @@
 import { Date, Document } from 'mongoose';
+
 import { Role } from '../types';
 
 export interface IUser extends Document {
@@ -8,4 +9,3 @@ export interface IUser extends Document {
   role: Role;
   createdAt: Date;
 }
-
