+++ src/schemas/User.ts
@@ -1,4 +1,5 @@
 import mongoose, { Schema } from 'mongoose';
+
 import { IUser } from '../interfaces/user';
 
 const UserSchema = new Schema<IUser>(
